system/requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.8
django-crispy-forms==1.13.0
django-dotenv==1.4.2
django-environ==0.8.1
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
h11==0.12.0
idna==3.3
outcome==1.1.0
prettytable==2.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.21
pyOpenSSL==21.0.0
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wcwidth==0.2.5
whitenoise==5.3.0
wsproto==1.0.0
